// models/user.ts
export interface User {
  email: string;
  id: number;
  password: string;
}

export interface Person {
  id: string;
  name: string;
  nickname: string;
  birthday: string;
  age: string;
  contactno: string;
  collegeAddress: string;
  homeAddress: string;
  email: string;
  sponsor: string;
  batch: string;
  confirmed: boolean;
  paid: boolean;
  balance: string;
  merit: string;
  demerit: string;
  position: string;
  photoUrl: string;
  lupon: string;
  paymentProofUrl: string;
  acknowledged: string;
  degprog: string;
}